Add all files begins with __subsearch__
__subsearch__ is added to all files that is written/created by SubSearch
So no unintended user files get overwritten or deleted

+++ SubSearch/src/utilities/file_manager.py
@@ -75,7 +75,7 @@ def move_files(cwd_path: str, prefered_extension: str, extension: str):
 # remove .zips
 def clean_up(cwd_path: str, extension: str):
     for file in os.listdir(cwd_path):
-        if file.endswith(extension):
+        if file.startswith("__subsearch__") and file.endswith(extension):
             log.output(f"Removing: {file}")
             os.remove(file)
 
@@ -110,3 +110,4 @@ def get_hash(file_name: str):
 
     except IOError as err:
         return log.output(err)
+
